feat: refactor auth callback templates
