+++ internal/auth/server_test.go
@@ -6,21 +6,23 @@ import (
 	"net/http/httptest"
 	"strings"
 	"testing"
+
+	"github.com/audetic/godeploy/internal/auth/templates"
 )
 
 func TestHashParamCallbackHTML(t *testing.T) {
-	// Test that the HashParamCallbackHTML contains the necessary JavaScript functions
-	if !strings.Contains(HashParamCallbackHTML, "getHashParam") {
-		t.Error("HashParamCallbackHTML does not contain getHashParam function")
+	// Test that the template contains the necessary JavaScript functions
+	if !strings.Contains(templates.HashParamCallbackHTML, "getHashParam") {
+		t.Error("Template does not contain getHashParam function")
 	}
-	if !strings.Contains(HashParamCallbackHTML, "processAuth") {
-		t.Error("HashParamCallbackHTML does not contain processAuth function")
+	if !strings.Contains(templates.HashParamCallbackHTML, "processAuth") {
+		t.Error("Template does not contain processAuth function")
 	}
-	if !strings.Contains(HashParamCallbackHTML, "window.location.hash") {
-		t.Error("HashParamCallbackHTML does not contain code to access window.location.hash")
+	if !strings.Contains(templates.HashParamCallbackHTML, "window.location.hash") {
+		t.Error("Template does not contain code to access window.location.hash")
 	}
-	if !strings.Contains(HashParamCallbackHTML, "/callback?access_token=") {
-		t.Error("HashParamCallbackHTML does not contain redirect to callback endpoint")
+	if !strings.Contains(templates.HashParamCallbackHTML, "/callback?access_token=") {
+		t.Error("Template does not contain redirect to callback endpoint")
 	}
 }
 
@@ -34,10 +36,12 @@ func TestRootHandler(t *testing.T) {
 		if r.URL.Path == "/" {
 			w.Header().Set("Content-Type", "text/html")
 			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(HashParamCallbackHTML))
+			w.Write([]byte(templates.HashParamCallbackHTML))
 			return
 		}
-		http.NotFound(w, r)
+		w.Header().Set("Content-Type", "text/plain")
+		w.WriteHeader(http.StatusNotFound)
+		w.Write([]byte("404 page not found"))
 	})
 
 	// Call the handler
@@ -58,174 +62,133 @@ func TestRootHandler(t *testing.T) {
 		t.Errorf("Expected content type %s, got %s", "text/html", contentType)
 	}
 
-	// Check that the response contains the HashParamCallbackHTML
+	// Check that the response contains the expected content
 	if !strings.Contains(w.Body.String(), "Processing Authentication") {
 		t.Error("Response does not contain expected HTML content")
 	}
 }
 
 func TestCallbackHandler(t *testing.T) {
-	// Create a request to the callback path with an access_token
-	req := httptest.NewRequest("GET", "/callback?access_token=test-token", nil)
-	w := httptest.NewRecorder()
-
-	// Create a handler function that simulates the callback handler in StartLocalServer
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		token := r.URL.Query().Get("access_token")
-		if token == "" {
-			http.Error(w, "Access token not found in the callback URL", http.StatusBadRequest)
-			return
-		}
-
-		w.Header().Set("Content-Type", "text/html")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(SuccessHTML))
-	})
-
-	// Call the handler
-	handler.ServeHTTP(w, req)
-
-	// Check the response
-	resp := w.Result()
-	defer resp.Body.Close()
-
-	// Check the status code
-	if resp.StatusCode != http.StatusOK {
-		t.Errorf("Expected status code %d, got %d", http.StatusOK, resp.StatusCode)
+	tests := []struct {
+		name           string
+		url            string
+		expectedStatus int
+		expectedBody   string
+	}{
+		{
+			name:           "success with token",
+			url:            "/callback?access_token=test-token",
+			expectedStatus: http.StatusOK,
+			expectedBody:   "Authentication Successful",
+		},
+		{
+			name:           "missing token",
+			url:            "/callback",
+			expectedStatus: http.StatusBadRequest,
+			expectedBody:   "Authentication Failed",
+		},
+		{
+			name:           "error response",
+			url:            "/callback?error=access_denied&error_code=otp_expired&error_description=Email+link+is+invalid+or+has+expired",
+			expectedStatus: http.StatusBadRequest,
+			expectedBody:   "Authentication Failed",
+		},
 	}
 
-	// Check the content type
-	contentType := resp.Header.Get("Content-Type")
-	if contentType != "text/html" {
-		t.Errorf("Expected content type %s, got %s", "text/html", contentType)
-	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			req := httptest.NewRequest("GET", tt.url, nil)
+			w := httptest.NewRecorder()
 
-	// Check that the response contains the SuccessHTML
-	if !strings.Contains(w.Body.String(), "Authentication Successful") {
-		t.Error("Response does not contain expected HTML content")
-	}
-}
+			// Create a handler function that simulates the callback handler in StartLocalServer
+			handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+				// Get the access token from the query parameters
+				token := r.URL.Query().Get("access_token")
 
-func TestMissingTokenHandler(t *testing.T) {
-	// Create a request to the callback path without an access_token
-	req := httptest.NewRequest("GET", "/callback", nil)
-	w := httptest.NewRecorder()
+				// Check for error parameters
+				errorParam := r.URL.Query().Get("error")
+				errorCode := r.URL.Query().Get("error_code")
+				errorDescription := r.URL.Query().Get("error_description")
 
-	// Create a handler function that simulates the callback handler in StartLocalServer
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		token := r.URL.Query().Get("access_token")
-		if token == "" {
-			http.Error(w, "Access token not found in the callback URL", http.StatusBadRequest)
-			return
-		}
+				if errorParam != "" {
+					// Handle error case
+					errorMessage := "Authentication failed."
+					if errorDescription != "" {
+						errorMessage = errorDescription
+					}
 
-		w.Header().Set("Content-Type", "text/html")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(SuccessHTML))
-	})
+					w.Header().Set("Content-Type", "text/html")
+					w.WriteHeader(http.StatusBadRequest)
+					w.Write([]byte(fmt.Sprintf(templates.ErrorHTML, errorMessage, errorParam, errorCode)))
+					return
+				}
 
-	// Call the handler
-	handler.ServeHTTP(w, req)
+				if token == "" {
+					// Handle missing token case with HTML error page
+					w.Header().Set("Content-Type", "text/html")
+					w.WriteHeader(http.StatusBadRequest)
+					w.Write([]byte(fmt.Sprintf(templates.ErrorHTML, "Access token not found in the callback URL", "missing_token", "")))
+					return
+				}
 
-	// Check the response
-	resp := w.Result()
-	defer resp.Body.Close()
+				// Check if this is a direct callback with a hash fragment
+				if !strings.Contains(r.URL.RawQuery, "access_token") &&
+					!strings.Contains(r.URL.RawQuery, "error") {
+					w.Header().Set("Content-Type", "text/html")
+					w.WriteHeader(http.StatusOK)
+					w.Write([]byte(templates.HashParamCallbackHTML))
+					return
+				}
 
-	// Check the status code
-	if resp.StatusCode != http.StatusBadRequest {
-		t.Errorf("Expected status code %d, got %d", http.StatusBadRequest, resp.StatusCode)
-	}
-}
+				w.Header().Set("Content-Type", "text/html")
+				w.WriteHeader(http.StatusOK)
+				w.Write([]byte(templates.SuccessHTML))
+			})
 
-func TestErrorHandling(t *testing.T) {
-	// Create a request to the callback path with error parameters
-	req := httptest.NewRequest("GET", "/callback?error=access_denied&error_code=otp_expired&error_description=Email+link+is+invalid+or+has+expired", nil)
-	w := httptest.NewRecorder()
+			// Call the handler
+			handler.ServeHTTP(w, req)
 
-	// Create a handler function that simulates the callback handler in StartLocalServer
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// Check for error parameters
-		errorParam := r.URL.Query().Get("error")
-		errorCode := r.URL.Query().Get("error_code")
-		errorDescription := r.URL.Query().Get("error_description")
+			// Check the response
+			resp := w.Result()
+			defer resp.Body.Close()
 
-		if errorParam != "" {
-			// Handle error case
-			errorMessage := "Authentication failed."
-			if errorDescription != "" {
-				errorMessage = errorDescription
+			// Check the status code
+			if resp.StatusCode != tt.expectedStatus {
+				t.Errorf("Expected status code %d, got %d", tt.expectedStatus, resp.StatusCode)
 			}
 
-			// Return the error HTML
-			w.Header().Set("Content-Type", "text/html")
-			w.WriteHeader(http.StatusBadRequest)
-			w.Write([]byte(fmt.Sprintf(
-				ErrorHTML,
-				errorMessage,
-				errorParam,
-				errorCode,
-			)))
-			return
-		}
+			// Check the content type
+			contentType := resp.Header.Get("Content-Type")
+			if tt.expectedStatus == http.StatusOK || tt.expectedStatus == http.StatusBadRequest {
+				if contentType != "text/html" {
+					t.Errorf("Expected content type text/html, got %s", contentType)
+				}
+			} else {
+				if contentType != "text/plain" {
+					t.Errorf("Expected content type text/plain, got %s", contentType)
+				}
+			}
 
-		// If no error, check for token
-		token := r.URL.Query().Get("access_token")
-		if token == "" {
-			http.Error(w, "Access token not found in the callback URL", http.StatusBadRequest)
-			return
-		}
-
-		w.Header().Set("Content-Type", "text/html")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(SuccessHTML))
-	})
-
-	// Call the handler
-	handler.ServeHTTP(w, req)
-
-	// Check the response
-	resp := w.Result()
-	defer resp.Body.Close()
-
-	// Check the status code
-	if resp.StatusCode != http.StatusBadRequest {
-		t.Errorf("Expected status code %d, got %d", http.StatusBadRequest, resp.StatusCode)
-	}
-
-	// Check the content type
-	contentType := resp.Header.Get("Content-Type")
-	if contentType != "text/html" {
-		t.Errorf("Expected content type %s, got %s", "text/html", contentType)
-	}
-
-	// Check that the response contains the error information
-	body := w.Body.String()
-	if !strings.Contains(body, "Authentication Failed") {
-		t.Error("Response does not contain expected error title")
-	}
-	if !strings.Contains(body, "Email link is invalid or has expired") {
-		t.Error("Response does not contain expected error description")
-	}
-	if !strings.Contains(body, "access_denied") {
-		t.Error("Response does not contain expected error type")
-	}
-	if !strings.Contains(body, "otp_expired") {
-		t.Error("Response does not contain expected error code")
+			// Check that the response contains the expected content
+			if !strings.Contains(w.Body.String(), tt.expectedBody) {
+				t.Errorf("Response does not contain expected content: %s", tt.expectedBody)
+			}
+		})
 	}
 }
 
 func TestHashParamErrorHandling(t *testing.T) {
-	// Test that the HashParamCallbackHTML handles error parameters correctly
-	if !strings.Contains(HashParamCallbackHTML, "getHashParam('error')") {
-		t.Error("HashParamCallbackHTML does not check for error parameter")
+	// Test that the template handles error parameters correctly
+	if !strings.Contains(templates.HashParamCallbackHTML, "getHashParam('error')") {
+		t.Error("Template does not check for error parameter")
 	}
-	if !strings.Contains(HashParamCallbackHTML, "getHashParam('error_code')") {
-		t.Error("HashParamCallbackHTML does not check for error_code parameter")
+	if !strings.Contains(templates.HashParamCallbackHTML, "getHashParam('error_code')") {
+		t.Error("Template does not check for error_code parameter")
 	}
-	if !strings.Contains(HashParamCallbackHTML, "getHashParam('error_description')") {
-		t.Error("HashParamCallbackHTML does not check for error_description parameter")
+	if !strings.Contains(templates.HashParamCallbackHTML, "getHashParam('error_description')") {
+		t.Error("Template does not check for error_description parameter")
 	}
-	if !strings.Contains(HashParamCallbackHTML, "'/callback?error='") {
-		t.Error("HashParamCallbackHTML does not contain redirect for error case")
+	if !strings.Contains(templates.HashParamCallbackHTML, "'/callback?error='") {
+		t.Error("Template does not contain redirect for error case")
 	}
 }
