+++ internal/auth/templates/error.html
@@ -1,0 +1,71 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>GoDeploy Authentication Error</title>
+    <style>
+      body {
+        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
+        background-color: #f5f5f5;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        height: 100vh;
+        margin: 0;
+      }
+      .container {
+        background-color: white;
+        border-radius: 8px;
+        box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+        padding: 40px;
+        text-align: center;
+        max-width: 500px;
+      }
+      h1 {
+        color: #e74c3c;
+        margin-bottom: 20px;
+      }
+      p {
+        color: #666;
+        line-height: 1.6;
+        margin-bottom: 30px;
+      }
+      .error-icon {
+        color: #e74c3c;
+        font-size: 64px;
+        margin-bottom: 20px;
+      }
+      .error-details {
+        background-color: #f9f9f9;
+        border-radius: 4px;
+        padding: 15px;
+        text-align: left;
+        margin-bottom: 20px;
+        font-family: monospace;
+        font-size: 14px;
+      }
+      .close-text {
+        font-size: 14px;
+        color: #999;
+        margin-top: 30px;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="container">
+      <div class="error-icon">✗</div>
+      <h1>Authentication Failed</h1>
+      <p>%s</p>
+      <div class="error-details">
+        <strong>Error:</strong> %s<br />
+        <strong>Code:</strong> %s
+      </div>
+      <p class="close-text">This window will automatically close in a few seconds...</p>
+    </div>
+    <script>
+      // Auto-close the window after 10 seconds
+      setTimeout(function () {
+        window.close();
+      }, 10000);
+    </script>
+  </body>
+</html>
