+++ internal/auth/templates/success.html
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>GoDeploy Authentication</title>
+    <style>
+      body {
+        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
+        background-color: #f5f5f5;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        height: 100vh;
+        margin: 0;
+      }
+      .container {
+        background-color: white;
+        border-radius: 8px;
+        box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+        padding: 40px;
+        text-align: center;
+        max-width: 500px;
+      }
+      h1 {
+        color: #333;
+        margin-bottom: 20px;
+      }
+      p {
+        color: #666;
+        line-height: 1.6;
+        margin-bottom: 30px;
+      }
+      .success-icon {
+        color: #4caf50;
+        font-size: 64px;
+        margin-bottom: 20px;
+      }
+      .close-text {
+        font-size: 14px;
+        color: #999;
+        margin-top: 30px;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="container">
+      <h1>Authentication Successful!</h1>
+      <p>You have successfully authenticated with GoDeploy. You can now close this window and return to the CLI.</p>
+      <p class="close-text">This window will automatically close in a few seconds...</p>
+    </div>
+    <script>
+      // Auto-close the window after 7 seconds
+      setTimeout(function () {
+        window.close();
+      }, 7000);
+    </script>
+  </body>
+</html>
